requirements.txt
djangorestframework==3.14.0
environs[django]==9.5.0
gunicorn==20.1.0
psycopg2==2.9.5
rollbar==0.16.3
Pillow==9.4.0
python-telegram-bot==13.15
django-phonenumber-field==7.0.2
phonenumbers==8.13.4
yookassa==2.3.3
django-widget-tweaks==1.4.12
Django~=4.1.5